Large swaths of refactoring: - Fixed many compiler warnings - Removed "AasHandle" and used references to the model adapter directly, instead

+++ Core/Systems/MapSector/SectorLoc.cs
@@ -41,7 +41,7 @@ public readonly struct SectorLoc
     [TempleDllLocation(0x100819c0)]
     public ulong Pack()
     {
-        return ((ulong) Y << 26) | (ulong) (X & 0x3FFFFFF);
+        return ((ulong) Y << 26) | (uint) (X & 0x3FFFFFF);
     }
 
     public static SectorLoc Unpack(ulong fieldValue)
